Build querybuilder before running it. Fixes #126.
