+++ frontend/webpack/querybuilder.config.js
@@ -1,0 +1,3 @@
+// @ts-check
+
+module.exports = require('./base').querybuilderConfig;
